refactor: 추상화 레벨에 맞지 않게 import해서 사용하고 있던 페이지 간 공통 컴포넌트 및 상수 분리 #249
