+++ frontend/src/components/FavoriteTab/FavoriteTab.styled.ts
