+++ frontend/src/components/TodoTab/TodoTab.test.tsx
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/named
 import { screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { describe, expect, it } from 'vitest';
