+++ frontend/src/components/TodoTab/index.tsx
@@ -1,18 +1,18 @@
-import TodoEditItem from '../TodoEditItem';
 import * as S from './TodoTab.styled';
-import { TODO_TAB_TEXT } from '../../EditPage.constants';
-import useAddTodoBottomSheet from '../../hooks/useAddTodoBottomSheet';
-import useDeleteTodoMutation from '../../hooks/useDeleteTodoMutation';
-import useEditTodoBottomSheet from '../../hooks/useEditTodoBottomSheet';
+import TodoEditItem from '../TodoEditItem';
+import useRouteTodoQuery from './hooks/useRouteTodoQuery';
+import { TODO_TAB_TEXT } from './TodoTab.constants';
 
 import IconButton from '@/components/Button/IconButton';
 import Icon from '@/components/Icon';
 import { MAX_TODO_ITEM_LENGTH, TODO_TYPE } from '@/constants/config';
 import { TODO_TOAST_MESSAGE } from '@/constants/message';
+import useAddTodoBottomSheet from '@/hooks/todo/useAddTodoBottomSheet';
+import useDeleteTodoMutation from '@/hooks/todo/useDeleteTodoMutation';
+import useEditTodoBottomSheet from '@/hooks/todo/useEditTodoBottomSheet';
+import useTodoListQuery from '@/hooks/todo/useTodoListQuery';
 import useQueryParamsDate from '@/hooks/useQueryParamsDate';
-import useRouteTodoQuery from '@/hooks/useRouteTodoQuery';
 import useToast from '@/hooks/useToast';
-import useTodoListQuery from '@/hooks/useTodoListQuery';
 import { TodoType } from '@/types/todo';
 
 interface TodoTabProps {
