+++ frontend/src/mocks/handlers/todo.ts
@@ -47,7 +47,8 @@ const getNewTodo = (todoId: number) => {
 };
 
 const getTodayTodoHandler = () => {
-  return HttpResponse.json(TODAY_TODO);
+  return new HttpResponse(JSON.stringify({ errorCode: 'TODO_NOT_FOUND' }), { status: 404 });
+  // return HttpResponse.json(TODAY_TODO);
 };
 
 const getTomorrowTodoHandler = () => {
