+++ frontend/src/pages/MainPage/components/TodoListContainer/index.tsx
@@ -4,8 +4,8 @@ import * as S from './TodoListContainer.styled';
 
 import Icon from '@/components/Icon';
 import { ICON_MAPPER } from '@/constants/config';
+import useTodoListQuery from '@/hooks/todo/useTodoListQuery';
 import useQueryParamsDate from '@/hooks/useQueryParamsDate';
-import useTodoListQuery from '@/hooks/useTodoListQuery';
 import { colors } from '@/styles/theme';
 
 const TODO_CONTENT_MESSAGE = {
